requirements.txt
flask==3.1.2
numpy
pydub==0.25.1
soundfile
trafilatura==2.0.0
kokoro-onnx==0.4.9
marker-pdf==1.10.1
pytest==8.3.4
misaki[en]
